requirements.txt
pandas
python-pptx
streamlit
openai
